*Bug fix* Flyway issue
In default, flyway ever runs before hibernate
because the flyway try to run all migrations (
like DDL to create database tables), etc.

But I'm not using flyway to create database
tables, I'm using to create a user or execute a
script to remove some data for the database.

I has configured the flyway to run after the
hibernate DLL is executed creating a
@Component called FlywayConfiguration.kt, this
call the flyway to run all "migrations" in
db/migration folder manually.